리덕스 액션 타입 이름 변경, common 컴포넌트 최적화
