+++ next_reservation/__tests__/components/Auth/SignInModal.test.tsx
@@ -7,7 +7,7 @@ import {
   mockStoreValue,
   useDispatchMock,
   useSelectorMock,
-} from '../../../__mocks__/auth/authMocks';
+} from '../../../__mocks__/redux/reduxStateMocks';
 import userEvent from '@testing-library/user-event';
 
 const store = useMockStore;
@@ -15,11 +15,7 @@ const store = useMockStore;
 beforeEach(() => {
   const dispatchMock = useDispatchMock;
   store.dispatch = dispatchMock;
-  useSelectorMock.mockImplementation((selector) =>
-    selector({
-      user: mockStoreValue,
-    })
-  );
+  useSelectorMock.mockImplementation((selector) => selector(mockStoreValue));
 });
 
 afterEach(() => {
