+++ next_reservation/components/common/Selector.tsx
@@ -30,4 +30,4 @@ const Selector: React.FC<IProps> = ({
   );
 };
 
-export default Selector;
+export default React.memo(Selector);
